Begin to separate REST behavior from other behavior
