+++ boxsdk/object/api_json_object.py
@@ -1,0 +1,9 @@
+# coding: utf-8
+
+from __future__ import unicode_literals
+
+from boxsdk.object.base_api_json_object import BaseAPIJSONObject
+
+
+class APIJSONObject(BaseAPIJSONObject):
+    pass
